// src/app/core/services/model/emailConfiguration.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EmailConfiguration { 
    host?: string;
    port?: number;
    username?: string;
    password?: string;
    from?: string;
    attchmentMaxBytes?: number;
}