// src/app/core/services/model/systemConfigDto.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ControllerConfiguration } from './controllerConfiguration';
import { EmailConfiguration } from './emailConfiguration';
import { EthernetConfigurationInput } from './ethernetConfigurationInput';
import { SftpConfiguration } from './sftpConfiguration';
import { TimeConfiguration } from './timeConfiguration';

export interface SystemConfigDto { 
    ethernetConfigurationEth0?: EthernetConfigurationInput;
    ethernetConfigurationEth1?: EthernetConfigurationInput;
    controllerConfiguration?: ControllerConfiguration;
    timeConfiguration?: TimeConfiguration;
    emailConfiguration?: EmailConfiguration;
    sftpConfiguration?: SftpConfiguration;
    /**
     * german or english
     */
    language?: string;
}